sub version 024 Laser OK.

+++ source/Marlin/src/gcode/motion/M290.cpp
@@ -97,7 +97,7 @@ void GcodeSuite::M290() {
       #if ENABLED(TENLOG_TOUCH_LCD)
         char cmd[32];
         tl_Z_HOME_POS -= offs;
-        int16_t zoffset = tl_Z_HOME_POS * 1000;
+        int32_t zoffset = tl_Z_HOME_POS * 1000;
         sprintf_P(cmd, PSTR("setting.xZOffset.val=%d"), -zoffset);
         TLSTJC_println(cmd);
         sprintf_P(cmd, PSTR("babystep.xBaby.val=%d"), -zoffset);
